Modifying data seeders, adding a `Chat Group Feed Entry` client-side component,

+++ server/Chatify.Infrastructure/Data/Models/ChatMessage.cs
@@ -1,7 +1,7 @@
 ﻿using Cassandra.Mapping.Attributes;
 using Chatify.Application.Common.Mappings;
 using Metadata = System.Collections.Generic.IDictionary<string, string>;
-using ReactionCounts = System.Collections.Generic.IDictionary<int, long>;
+using ReactionCounts = System.Collections.Generic.IDictionary<short, long>;
 
 namespace Chatify.Infrastructure.Data.Models;
 
@@ -22,7 +22,7 @@ public class ChatMessage : IMapFrom<Domain.Entities.ChatMessage>
 
     public Metadata Metadata { get; set; } = new Dictionary<string, string>();
 
-    public ReactionCounts ReactionCounts { get; set; } = new Dictionary<int, long>();
+    public ReactionCounts ReactionCounts { get; set; } = new Dictionary<short, long>();
 
     public DateTimeOffset CreatedAt { get; set; }
